Create README.md
Detects dust particles in the air using Arduino and sensors attached to it. The voltage reading from the dust is then sent to the web application where there is an animation showing the concentration of dust particles in the air as well as playing a sound to correspond to that.